+++ VAST/core/submodels/actuation_submodels/pitching_actuation_model.py
@@ -6,11 +6,24 @@ from numpy.core.fromnumeric import size
 class EelActuationModel(csdl.Model):
     """
     Compute the mesh at each time step for the eel actuation model given the kinematic variables.
-    parameters
+    # the geometry and kinematics of the eel
+    # is define from https://www.cse-lab.ethz.ch/wp-content/papercite-data/pdf/kern2006b.pdf
+
+    Inputs
     ----------
     tail_amplitude : csdl variable [1,]
     tail_frequency : csdl variable [1,]
-    v_x : csdl variable [1,]
+    wave_number : csdl variable [1,]
+    linear_relation : csdl variable [1,]
+
+    Parameters
+    ----------
+    surface_names : list
+    surface_shapes : list
+    n_period : int
+    s_1_ind : int
+    s_2_ind : int
+
     Returns
     -------
     1. mesh[num_nodes,num_pts_chord, num_pts_span, 3] : csdl array
@@ -19,7 +32,6 @@ class EelActuationModel(csdl.Model):
     def initialize(self):
         self.parameters.declare('surface_names', types=list)
         self.parameters.declare('surface_shapes', types=list)
-        self.parameters.declare('mesh_unit', default='m')
         self.parameters.declare('n_period')
         self.parameters.declare('s_1_ind',default=3)
         self.parameters.declare('s_2_ind',default=None)
@@ -28,7 +40,6 @@ class EelActuationModel(csdl.Model):
     def define(self):
         surface_names = self.parameters['surface_names']
         surface_shapes = self.parameters['surface_shapes']
-        mesh_unit = self.parameters['mesh_unit']
         N_period = self.parameters['n_period']
         s_1_ind = self.parameters['s_1_ind'] # head region
         s_2_ind = self.parameters['s_2_ind'] # tail region
@@ -39,76 +50,80 @@ class EelActuationModel(csdl.Model):
         num_nodes = surface_shapes[0][0]
 
         ########################################################
-        # Defining kinematic variables for the fish
+        # 1. Defining kinematic variables for the fish
         ########################################################
 
-        v_inf = 0.4
-        lambda_ = 1
-        st = 0.15
-        A = 0.125
-        f = st*v_inf/A 
-        omg = 2*np.pi*f
+        tail_amplitude = self.declare_variable('tail_amplitude')
+        tail_frequency = self.declare_variable('tail_frequency')
+        wave_number = self.declare_variable('wave_number')
+        linear_relation = self.declare_variable('linear_relation')
 
-        tail_amplitude = self.declare_variable('tail_amplitude', val=A)
-        tail_frequency = self.declare_variable('tail_frequency', val=f)
-        wave_number = self.declare_variable('wave_number', val=lambda_)
-        linear_relation = self.declare_variable('linear_relation', val=0.03125)
+        omg = 2*np.pi*tail_frequency
         
-        self.print_var(tail_amplitude)
-        self.print_var(tail_frequency)
-        self.print_var(wave_number)
-
         t_temp_val = np.linspace(0, N_period, num_nodes)
         t_temp = self.create_input('t_temp', val = t_temp_val)
-        # print()
 
         t = t_temp/csdl.expand(tail_frequency,shape=t_temp.shape)
+        ########################################################
+
+        ########################################################
+        # 2. Defining the mesh of the fish
         ########################################################
 
         for i in range(num_surface):
             nx = surface_shapes[i][1]
             ny = surface_shapes[i][2]    
-            # define the x (length) discretization of the mesh
-            L = 1.0
-            # s_1_ind = 5 # head region
-            # s_2_ind = int(nx-3) # tail region
-            # s_1_ind = 3 # head region
-            # s_2_ind = int(nx-2) # tail region            
-            s1 = 0.04 * L
-            s2 = 0.95 * L
-            x_1 = (1-np.cos(np.linspace(0, np.pi/2,s_1_ind,endpoint=False)))/1*s1
-            x_2 = np.linspace(s1, s2, int(s_2_ind-s_1_ind),endpoint=False)
+
+            L = 1.0 # length of the fish
+            s1 = 0.04 * L # head region
+            s2 = 0.95 * L # tail region  
+
+            # define the x (length) discretization of the mesh:
+            # cos spacing for head
+            x_1 = (1-np.cos(np.linspace(0, np.pi/2,s_1_ind,endpoint=False)))*s1 
+            # linear spacing for body and tail
+            x_2 = np.linspace(s1, s2, int(s_2_ind-s_1_ind),endpoint=False) 
             x_3 = np.linspace(s2, L, int(nx-s_2_ind))
-            x = np.concatenate((x_1,x_2,x_3))
+            x = np.concatenate((x_1,x_2,x_3)) # size = nx
                         
             # define the x and y coordinates of the mesh
-            tensor_x = np.outer(x, np.ones(ny)).reshape(nx,ny,1)
-            tensor_z  = np.outer(np.arange(ny), np.ones(nx)).T.reshape(nx,ny,1)/(ny-1) - 0.5
+            tensor_x = np.outer(x, np.ones(ny)).reshape(nx,ny,1) # x coordinate
+            # for the z coordinate, tensor_z is just a expanded array from -0.5 to 0.5,
+            # need to scale it to the actual height of the fish
+            tensor_z  = np.outer(np.arange(ny), np.ones(nx)).T.reshape(nx,ny,1)/(ny-1) - 0.5 
 
             tensor_x_exp = np.einsum('i,jkl->ijkl',np.ones(num_nodes),tensor_x)
             tensor_z_exp = np.einsum('i,jkl->ijkl',np.ones(num_nodes),tensor_z)
 
-            tensor_x_csdl = self.declare_variable('mesh_x',val=tensor_x_exp)
+            x_exp = self.declare_variable('mesh_x',val=tensor_x_exp)
             tensor_z_csdl = self.declare_variable('mesh_y',val=tensor_z_exp)
 
-            a = 0.55*L
-            b = 0.08*L
-
+            a = 0.51*L
+            b = 0.08*L 
+            # from https://www.cse-lab.ethz.ch/wp-content/papercite-data/pdf/kern2006b.pdf to define the height of the fish
+            # h(x) = b (1 - ((x-a)/a)^2 )^0.5
             height = np.einsum('ik,j->ijk',np.ones((num_nodes,ny)),b*(1-((x-a)/a)**2)**0.5).reshape(num_nodes,nx,ny,1)
 
             t_exp = csdl.expand(t, shape=(num_nodes,nx,ny,1),indices='i->ijkl')
             tail_amplitude_exp = csdl.expand(tail_amplitude, shape=(num_nodes,nx,ny,1),indices='i->ijkl')
             wave_number_exp = csdl.expand(wave_number, shape=(num_nodes,nx,ny,1),indices='i->ijkl')
             linear_relation_exp = csdl.expand(linear_relation, shape=(num_nodes,nx,ny,1),indices='i->ijkl') 
-
-            x_exp = tensor_x_csdl
+            omg_exp = csdl.expand(omg, shape=(num_nodes,nx,ny,1),indices='i->ijkl')
             
             # y = -A(x+0.03/1.03)*sin(2pix/lambda - omg*t)
             # ydot = -A(x+0.03/1.03)*cos(2pix/lambda - omg*t)*(-omg)
 
-            y = -tail_amplitude_exp*((x_exp+linear_relation_exp)/(linear_relation_exp+1)) * csdl.sin( - omg*t_exp)
-            y_dot =  -tail_amplitude_exp*((x_exp+linear_relation_exp)/(linear_relation_exp+1))*csdl.cos( - omg*t_exp)*(-omg)
+            # y is the lateral displacement of the fish
+            # y = 0.125 * (s+0.03125)/1.03125 * sin(2*pi*x/lambda - omg*t)
+            # ydot = 0.125 * (s+0.03125)/1.03125 * cos(2*pi*x/lambda - omg*t) * (-omg)
 
+            # y = tail_amplitude_exp*((x_exp+linear_relation_exp)/(linear_relation_exp+1)) * csdl.sin(np.pi*2*x_exp/wave_number_exp - omg_exp*t_exp)
+            # y_dot =  tail_amplitude_exp*((x_exp+linear_relation_exp)/(linear_relation_exp+1))*csdl.cos(np.pi*2*x_exp/wave_number_exp - omg_exp*t_exp)*(-omg_exp)
+
+            y = tail_amplitude_exp*((x_exp+linear_relation_exp)/(linear_relation_exp+1)) * csdl.sin( - omg_exp*t_exp)
+            y_dot =  tail_amplitude_exp*((x_exp+linear_relation_exp)/(linear_relation_exp+1))*csdl.cos( - omg_exp*t_exp)*(-omg_exp)
+
+            # velocity of the fish on its collocation points
             coll_vel = self.create_output(name=surface_names[i]+'_coll_vel',val=np.zeros((num_nodes,nx-1,ny-1,3)))
             
             coll_vel[:,:,:,1] = (y_dot[:,:-1,:-1,:]+y_dot[:,1:,:-1,:]+y_dot[:,:-1,1:,:]+y_dot[:,1:,1:,:])/4 
@@ -116,12 +131,10 @@ class EelActuationModel(csdl.Model):
             mesh = self.create_output(surface_names[i],val=np.zeros((num_nodes,nx,ny,3)))
 
 
-            mesh[:,:,:,0] = tensor_x_csdl    
+            mesh[:,:,:,0] = x_exp    
             mesh[:,:,:,1] = y
-            mesh[:,:,:,2] = tensor_z_csdl * height * 2
-
-            self.register_output('y_lateral', y[0:,0:,0,0])
-
+            mesh[:,:,:,2] = tensor_z_csdl * height * 2 
+            # this is due to the fact that the original height is from -0.5 to 0.5, need to scale it to the actual height of the fish
 
 
 if __name__ == "__main__":
